+++ tests/testthat/test-RushWorker.R
@@ -35,9 +35,13 @@ test_that("a worker is registered", {
 
   worker_info = rush$worker_info
   expect_data_table(worker_info, nrows = 1)
-  expect_names(names(worker_info), permutation.of = c("worker_id", "pid", "hostname", "heartbeat", "state"))
+  expect_names(
+    names(worker_info),
+    permutation.of = c("worker_id", "pid", "hostname", "profile", "heartbeat", "state")
+  )
   expect_equal(worker_info$worker_id, rush$worker_id)
   expect_equal(worker_info$pid, Sys.getpid())
+  expect_true(is.na(worker_info$profile))
   expect_equal(rush$worker_ids, rush$worker_id)
   expect_equal(rush$worker_info$state, "running")
 })